test: 테스트 코드 깨짐 수정
- DB Cleaser 추가
- Lock 의 경우 Transactional 내에서 수행되도록 수정
- 동시성 테스트 클래스 분리
