+++ src/test/java/kr/hhplus/be/server/interfaces/coupon/CouponControllerIntegrationTest.java
@@ -7,6 +7,7 @@ import kr.hhplus.be.server.fixtures.CouponFixtures;
 import kr.hhplus.be.server.fixtures.UserFixtures;
 import kr.hhplus.be.server.infrastructure.coupon.CouponIssueJpaRepository;
 import kr.hhplus.be.server.infrastructure.coupon.CouponJpaRepository;
+import kr.hhplus.be.server.infrastructure.support.DatabaseCleanup;
 import kr.hhplus.be.server.infrastructure.user.UserJpaRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
@@ -45,10 +46,12 @@ class CouponControllerIntegrationTest {
     @Autowired
     private UserJpaRepository userJpaRepository;
 
+    @Autowired
+    private DatabaseCleanup databaseCleanup;
+
     @BeforeEach
     void setUp() {
-        couponIssueJpaRepository.deleteAll();
-        couponJpaRepository.deleteAll();
+        databaseCleanup.truncateAllTables();
     }
 
     @Nested
